+++ pennylane_lightning/core/src/simulators/lightning_gpu/StateVectorCudaManaged.hpp
@@ -307,11 +307,11 @@ class StateVectorCudaManaged
             if (adjoint) {
                 auto rot_matrix =
                     cuGates::getRot<CFP_t>(params[2], params[1], params[0]);
-                applyDeviceMatrixGate(rot_matrix.data(), ctrls, tgts, true);
+                applyDeviceMatrixGate_(rot_matrix.data(), ctrls, tgts, true);
             } else {
                 auto rot_matrix =
                     cuGates::getRot<CFP_t>(params[0], params[1], params[2]);
-                applyDeviceMatrixGate(rot_matrix.data(), ctrls, tgts, false);
+                applyDeviceMatrixGate_(rot_matrix.data(), ctrls, tgts, false);
             }
         } else if (opName == "Matrix") {
             DataBuffer<CFP_t, int> d_matrix{
@@ -324,8 +324,8 @@ class StateVectorCudaManaged
                                                        ctrls.rend()};
             const std::vector<std::size_t> tgts_local{tgts.rbegin(),
                                                       tgts.rend()};
-            applyDeviceMatrixGate(d_matrix.getData(), ctrls_local, tgts_local,
-                                  adjoint);
+            applyDeviceMatrixGate_(d_matrix.getData(), ctrls_local, tgts_local,
+                                   adjoint);
         } else if (par_gates_.find(opName) != par_gates_.end()) {
             const std::vector<std::size_t> wires_local{wires.rbegin(),
                                                        wires.rend()};
@@ -341,12 +341,13 @@ class StateVectorCudaManaged
 
             if (!gate_cache_.gateExists(opName, par[0]) &&
                 gate_matrix.empty()) {
-                std::string message = "Currently unsupported gate: " + opName;
+                std::string message = "Currently unsupported gate: " + opName +
+                                      "and no matrix is provided.";
                 throw LightningException(message);
             } else if (!gate_cache_.gateExists(opName, par[0])) {
                 gate_cache_.add_gate(opName, par[0], gate_matrix);
             }
-            applyDeviceMatrixGate(
+            applyDeviceMatrixGate_(
                 gate_cache_.get_gate_device_ptr(opName, par[0]), ctrls_local,
                 tgts_local, adjoint);
         }
@@ -361,19 +362,22 @@ class StateVectorCudaManaged
      * @param tgt_wires Wires to apply gate to.
      * @param adjoint Indicates whether to use adjoint of gate.
      * @param params Optional parameter list for parametric gates.
-     * @param gate_matrix Optional std gate matrix if opName doesn't exist.
+     * @param gate_matrix Optional gate matrix vector on the host if opName
+     * doesn't exist.
      */
-    void applyOperation(
-        const std::string &opName,
-        const std::vector<std::size_t> &controlled_wires,
-        const std::vector<bool> &controlled_values,
-        const std::vector<std::size_t> &tgt_wires, bool adjoint = false,
-        const std::vector<Precision> &params = {0.0},
-        [[maybe_unused]] const std::vector<ComplexT> &gate_matrix = {}) {
-        PL_ABORT_IF_NOT(opName == "GlobalPhase",
-                        "Only GlobalPhase gate is supported.");
+    void applyOperation(const std::string &opName,
+                        const std::vector<std::size_t> &controlled_wires,
+                        const std::vector<bool> &controlled_values,
+                        const std::vector<std::size_t> &tgt_wires,
+                        bool adjoint = false,
+                        const std::vector<Precision> &params = {0.0},
+                        const std::vector<ComplexT> &gate_matrix = {}) {
+        PL_ABORT_IF_NOT(
+            areVecsDisjoint<std::size_t>(controlled_wires, tgt_wires),
+            "`controlled_wires` and target wires must be disjoint.");
         PL_ABORT_IF(controlled_wires.size() != controlled_values.size(),
-                    "`ctrls` and `ctrls_values` must have the same size.");
+                    "`controlled_wires` and `controlled_values` must have the "
+                    "same size.");
         auto ctrlsInt = NormalizeCastIndices<std::size_t, int>(
             controlled_wires, BaseType::getNumQubits());
         auto tgtsInt = NormalizeCastIndices<std::size_t, int>(
@@ -381,11 +385,103 @@ class StateVectorCudaManaged
         auto ctrls_valuesInt =
             Pennylane::Util::cast_vector<bool, int>(controlled_values);
 
-        if (opName == "GlobalPhase") {
-            const std::vector<std::string> names(tgt_wires.size(), "I");
+        if (opName == "MultiRZ") {
+            const std::vector<std::string> names(tgtsInt.size(), {"RZ"});
+            applyParametricPauliGeneralGate_(names, ctrlsInt, ctrls_valuesInt,
+                                             tgtsInt, params.front(), adjoint);
+        } else if (opName == "GlobalPhase") {
+            const std::vector<std::string> names(tgtsInt.size(), "I");
             applyParametricPauliGeneralGate_(names, ctrlsInt, ctrls_valuesInt,
                                              tgtsInt, 2 * params[0], adjoint);
+
+        } else if (native_gates_.find(opName) != native_gates_.end()) {
+            applyParametricPauliGeneralGate_({opName}, ctrlsInt,
+                                             ctrls_valuesInt, tgtsInt,
+                                             params.front(), adjoint);
+        } else if (opName == "Rot") {
+            auto rot_matrix =
+                adjoint
+                    ? cuGates::getRot<CFP_t>(params[2], params[1], params[0])
+                    : cuGates::getRot<CFP_t>(params[0], params[1], params[2]);
+            applyDeviceGeneralGate_(rot_matrix.data(), ctrlsInt, tgtsInt,
+                                    ctrls_valuesInt, adjoint);
+        } else if (par_gates_.find(opName) != par_gates_.end()) {
+            // TODO: offload to par_gates_ if available
+            auto &gateMap =
+                cuGates::DynamicGateDataAccess<PrecisionT>::getInstance();
+            auto &&matrix_cu = gateMap.getGateData(opName, params);
+
+            gate_cache_.add_gate(opName, params[0], matrix_cu);
+
+            applyDeviceGeneralGate_(
+                gate_cache_.get_gate_device_ptr(opName, params[0]), ctrlsInt,
+                tgtsInt, ctrls_valuesInt, adjoint);
+        } else { // No offloadable function call; defer to matrix passing
+            auto &&par =
+                (params.empty()) ? std::vector<Precision>{0.0} : params;
+            if (!gate_cache_.gateExists(opName, par[0]) &&
+                gate_matrix.empty()) {
+                std::string message = "Currently unsupported gate: " + opName +
+                                      " and no matrix provided.";
+                throw LightningException(message);
+            } else if (!gate_cache_.gateExists(opName, par[0])) {
+                std::vector<CFP_t> matrix_cu(gate_matrix.size());
+                std::transform(
+                    gate_matrix.begin(), gate_matrix.end(), matrix_cu.begin(),
+                    [](const std::complex<Precision> &x) {
+                        return cuUtil::complexToCu<std::complex<Precision>>(x);
+                    });
+
+                gate_cache_.add_gate(opName, par[0], matrix_cu);
+            }
+            std::reverse(ctrlsInt.begin(), ctrlsInt.end());
+            std::reverse(tgtsInt.begin(), tgtsInt.end());
+            std::reverse(ctrls_valuesInt.begin(), ctrls_valuesInt.end());
+            applyDeviceGeneralGate_(
+                gate_cache_.get_gate_device_ptr(opName, par[0]), ctrlsInt,
+                tgtsInt, ctrls_valuesInt, adjoint);
         }
+    }
+
+    /**
+     * @brief Apply a single gate to the state vector.
+     *
+     * @param gate_matrix Gate matrix data (in row-major format).
+     * @param matrix_size Size of the matrix.
+     * @param controlled_wires Control wires.
+     * @param controlled_values Control values (false or true).
+     * @param tgt_wires Target wires to apply gate to.
+     * @param inverse Indicates whether to use adjoint of gate.
+     */
+    void applyControlledMatrix(const ComplexT *gate_matrix,
+                               const std::size_t matrix_size,
+                               const std::vector<std::size_t> &controlled_wires,
+                               const std::vector<bool> &controlled_values,
+                               const std::vector<std::size_t> &tgt_wires,
+                               bool inverse = false) {
+        PL_ABORT_IF_NOT(
+            areVecsDisjoint<std::size_t>(controlled_wires, tgt_wires),
+            "`controlled_wires` and `target wires` must be disjoint.");
+        PL_ABORT_IF(controlled_wires.size() != controlled_values.size(),
+                    "`controlled_wires` and `controlled_values` must have the "
+                    "same size.");
+        DataBuffer<CFP_t, int> d_matrix{
+            matrix_size, BaseType::getDataBuffer().getDevTag(), true};
+        d_matrix.CopyHostDataToGpu(gate_matrix, matrix_size, false);
+        // ensure wire indexing correctly preserved for tensor-observables
+        auto ctrlsInt = NormalizeCastIndices<std::size_t, int>(
+            controlled_wires, BaseType::getNumQubits());
+        auto tgtsInt = NormalizeCastIndices<std::size_t, int>(
+            tgt_wires, BaseType::getNumQubits());
+        auto ctrls_valuesInt =
+            Pennylane::Util::cast_vector<bool, int>(controlled_values);
+
+        std::reverse(ctrlsInt.begin(), ctrlsInt.end());
+        std::reverse(tgtsInt.begin(), tgtsInt.end());
+        std::reverse(ctrls_valuesInt.begin(), ctrls_valuesInt.end());
+
+        applyDeviceGeneralGate_(d_matrix.getData(), ctrlsInt, tgtsInt,
+                                ctrls_valuesInt, inverse);
     }
 
     /**
@@ -510,47 +606,47 @@ class StateVectorCudaManaged
                             bool adjoint) {
         static const std::string name{"PauliX"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyPauliY(const std::vector<std::size_t> &wires,
                             bool adjoint) {
         static const std::string name{"PauliY"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyPauliZ(const std::vector<std::size_t> &wires,
                             bool adjoint) {
         static const std::string name{"PauliZ"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyHadamard(const std::vector<std::size_t> &wires,
                               bool adjoint) {
         static const std::string name{"Hadamard"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyS(const std::vector<std::size_t> &wires, bool adjoint) {
         static const std::string name{"S"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyT(const std::vector<std::size_t> &wires, bool adjoint) {
         static const std::string name{"T"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyRX(const std::vector<std::size_t> &wires, bool adjoint,
                         Precision param) {
@@ -588,38 +684,38 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(gate_key,
                                  cuGates::getPhaseShift<CFP_t>(param));
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
 
     /* two-qubit gates */
     inline void applyCNOT(const std::vector<std::size_t> &wires, bool adjoint) {
         static const std::string name{"CNOT"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyCY(const std::vector<std::size_t> &wires, bool adjoint) {
         static const std::string name{"CY"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyCZ(const std::vector<std::size_t> &wires, bool adjoint) {
         static const std::string name{"CZ"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applySWAP(const std::vector<std::size_t> &wires, bool adjoint) {
         static const std::string name{"SWAP"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param), {},
+                               wires, adjoint);
     }
     inline void applyIsingXX(const std::vector<std::size_t> &wires,
                              bool adjoint, Precision param) {
@@ -643,8 +739,8 @@ class StateVectorCudaManaged
         if (!gate_cache_.gateExists(gate_key)) {
             gate_cache_.add_gate(gate_key, cuGates::getIsingXY<CFP_t>(param));
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
     }
     inline void applyCRot(const std::vector<std::size_t> &wires, bool adjoint,
                           const std::vector<Precision> &params) {
@@ -682,8 +778,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(gate_key,
                                  cuGates::getSingleExcitation<CFP_t>(param));
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
     }
     inline void
     applySingleExcitationMinus(const std::vector<std::size_t> &wires,
@@ -694,8 +790,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getSingleExcitationMinus<CFP_t>(param));
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
     }
     inline void applySingleExcitationPlus(const std::vector<std::size_t> &wires,
                                           bool adjoint, Precision param) {
@@ -705,8 +801,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getSingleExcitationPlus<CFP_t>(param));
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
     }
 
     /* three-qubit gates */
@@ -714,35 +810,35 @@ class StateVectorCudaManaged
                              bool adjoint) {
         static const std::string name{"Toffoli"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.begin(), wires.end() - 1}, {wires.back()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.begin(), wires.end() - 1}, {wires.back()},
+                               adjoint);
     }
     inline void applyCSWAP(const std::vector<std::size_t> &wires,
                            bool adjoint) {
         static const std::string name{"SWAP"};
         static const Precision param = 0.0;
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                              {wires.front()}, {wires.begin() + 1, wires.end()},
-                              adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                               {wires.front()},
+                               {wires.begin() + 1, wires.end()}, adjoint);
     }
 
     /* four-qubit gates */
     inline void applyDoubleExcitation(const std::vector<std::size_t> &wires,
                                       bool adjoint, Precision param) {
         auto &&mat = cuGates::getDoubleExcitation<CFP_t>(param);
-        applyDeviceMatrixGate(mat.data(), {}, wires, adjoint);
+        applyDeviceMatrixGate_(mat.data(), {}, wires, adjoint);
     }
     inline void
     applyDoubleExcitationMinus(const std::vector<std::size_t> &wires,
                                bool adjoint, Precision param) {
         auto &&mat = cuGates::getDoubleExcitationMinus<CFP_t>(param);
-        applyDeviceMatrixGate(mat.data(), {}, wires, adjoint);
+        applyDeviceMatrixGate_(mat.data(), {}, wires, adjoint);
     }
     inline void applyDoubleExcitationPlus(const std::vector<std::size_t> &wires,
                                           bool adjoint, Precision param) {
         auto &&mat = cuGates::getDoubleExcitationPlus<CFP_t>(param);
-        applyDeviceMatrixGate(mat.data(), {}, wires, adjoint);
+        applyDeviceMatrixGate_(mat.data(), {}, wires, adjoint);
     }
 
     /* Multi-qubit gates */
@@ -815,8 +911,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(gate_key,
                                  cuGates::getGeneratorIsingXX<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
     inline PrecisionT
@@ -828,8 +924,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(gate_key,
                                  cuGates::getGeneratorIsingYY<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
     inline PrecisionT
@@ -841,8 +937,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(gate_key,
                                  cuGates::getGeneratorIsingZZ<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
 
@@ -855,8 +951,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(gate_key,
                                  cuGates::getGeneratorIsingXY<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return static_cast<PrecisionT>(0.5);
     }
 
@@ -943,8 +1039,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getGeneratorSingleExcitation<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
     inline PrecisionT
@@ -957,8 +1053,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getGeneratorSingleExcitationMinus<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
     inline PrecisionT
@@ -971,8 +1067,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getGeneratorSingleExcitationPlus<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
 
@@ -986,8 +1082,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getGeneratorDoubleExcitation<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
     inline PrecisionT
@@ -1000,8 +1096,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getGeneratorDoubleExcitationMinus<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
     inline PrecisionT
@@ -1014,8 +1110,8 @@ class StateVectorCudaManaged
             gate_cache_.add_gate(
                 gate_key, cuGates::getGeneratorDoubleExcitationPlus<CFP_t>());
         }
-        applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(gate_key), {},
-                              wires, adjoint);
+        applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(gate_key), {},
+                               wires, adjoint);
         return -static_cast<PrecisionT>(0.5);
     }
 
@@ -1024,8 +1120,8 @@ class StateVectorCudaManaged
         static const std::string name{"PauliZ"};
         static const Precision param = 0.0;
         for (const auto &w : wires) {
-            applyDeviceMatrixGate(gate_cache_.get_gate_device_ptr(name, param),
-                                  {}, {w}, adjoint);
+            applyDeviceMatrixGate_(gate_cache_.get_gate_device_ptr(name, param),
+                                   {}, {w}, adjoint);
         }
         return -static_cast<PrecisionT>(0.5);
     }
@@ -1499,25 +1595,41 @@ class StateVectorCudaManaged
      * @param tgts Target qubits.
      * @param use_adjoint Use adjoint of given gate.
      */
-    void applyDeviceMatrixGate(const CFP_t *matrix,
-                               const std::vector<std::size_t> &ctrls,
-                               const std::vector<std::size_t> &tgts,
-                               bool use_adjoint = false) {
+    void applyDeviceMatrixGate_(const CFP_t *matrix,
+                                const std::vector<std::size_t> &ctrls,
+                                const std::vector<std::size_t> &tgts,
+                                bool use_adjoint = false) {
+        auto ctrlsInt = NormalizeCastIndices<std::size_t, int>(
+            ctrls, BaseType::getNumQubits());
+        auto tgtsInt = NormalizeCastIndices<std::size_t, int>(
+            tgts, BaseType::getNumQubits());
+        auto ctrls_values = std::vector<int>(ctrls.size(), 1);
+
+        applyDeviceGeneralGate_(matrix, ctrlsInt, tgtsInt, ctrls_values,
+                                use_adjoint);
+    }
+
+    /**
+     * @brief Apply a given host or device-stored array representing the gate
+     * `matrix` to the state vector at qubit indices given by `tgts` and
+     * control-lines given by `ctrls`. The adjoint can be taken by setting
+     * `use_adjoint` to true.
+     *
+     * @param matrix Device data array in row-major order representing
+     * a given gate.
+     * @param ctrls Control line qubits.
+     * @param tgts Target qubits.
+     * @param ctrls_values Control values.
+     * @param use_adjoint Use adjoint of given gate. Defaults to false.
+     */
+    void applyDeviceGeneralGate_(const CFP_t *matrix,
+                                 const std::vector<int> &ctrls,
+                                 const std::vector<int> &tgts,
+                                 const std::vector<int> &ctrls_values,
+                                 bool use_adjoint = false) {
         void *extraWorkspace = nullptr;
         std::size_t extraWorkspaceSizeInBytes = 0;
         int nIndexBits = BaseType::getNumQubits();
-
-        std::vector<int> ctrlsInt(ctrls.size());
-        std::vector<int> tgtsInt(tgts.size());
-
-        std::transform(
-            ctrls.begin(), ctrls.end(), ctrlsInt.begin(), [&](std::size_t x) {
-                return static_cast<int>(BaseType::getNumQubits() - 1 - x);
-            });
-        std::transform(
-            tgts.begin(), tgts.end(), tgtsInt.begin(), [&](std::size_t x) {
-                return static_cast<int>(BaseType::getNumQubits() - 1 - x);
-            });
 
         cudaDataType_t data_type;
         custatevecComputeType_t compute_type;
@@ -1566,10 +1678,10 @@ class StateVectorCudaManaged
             /* cudaDataType_t */ data_type,
             /* custatevecMatrixLayout_t */ CUSTATEVEC_MATRIX_LAYOUT_ROW,
             /* const int32_t */ use_adjoint,
-            /* const int32_t* */ tgtsInt.data(),
+            /* const int32_t* */ tgts.data(),
             /* const uint32_t */ tgts.size(),
-            /* const int32_t* */ ctrlsInt.data(),
-            /* const int32_t* */ nullptr,
+            /* const int32_t* */ ctrls.data(),
+            /* const int32_t* */ ctrls_values.data(),
             /* const uint32_t */ ctrls.size(),
             /* custatevecComputeType_t */ compute_type,
             /* void* */ extraWorkspace,